requirements.txt
Flask
gunicorn
pydub
PyAudio
numpy
scipy
matplotlib
mysql-connector-python
psycopg2
